Moved templates to use new directory structure
