+++ src/generators/app/templates/_common/src/components/App/styles.scss
