+++ src/generators/app/templates/_common/src/components/ErrorBox/styles.css
