+++ src/generators/app/templates/basic/src/components/App/index.test.js
@@ -1,4 +1,4 @@
-const App = require('../App');
+const App = require('.');
 
 test('it renders', () => {
   const component = new App({ projectName: 'test-project' });
