+++ src/generators/app/templates/react/src/components/App/index.js
@@ -1,5 +1,5 @@
 const React = require('react');
-const styles = require('./App.scss');
+const styles = require('./styles.scss');
 const worm = require('./worm.svg');
 
 class App extends React.Component {
