+++ src/generators/app/templates/react/src/components/ErrorBox/index.js
@@ -1,5 +1,5 @@
 const React = require('react');
-const styles = require('./ErrorBox.css');
+const styles = require('./styles.css');
 
 class ErrorBox extends React.Component {
   componentDidMount() {
