+++ src/generators/component/templates/basic/component-with-d3.test.js
@@ -1,4 +1,4 @@
-const <%= className %> = require('../<%= className %>');
+const <%= className %> = require('.');
 
 test('it renders', () => {
   const component = new <%= className %>();
