+++ src/generators/component/templates/preact/component-with-d3.test.js
@@ -2,7 +2,7 @@ const { h } = require('preact');
 const render = require('preact-render-to-string');
 const htmlLooksLike = require('html-looks-like');
 
-const <%= className %> = require('../<%= className %>');
+const <%= className %> = require('.');
 
 describe('<%= className %>', () => {
   test('It renders', () => {
