+++ src/generators/component/templates/react/component.test.js
@@ -1,7 +1,7 @@
 const React = require('react');
 const renderer = require('react-test-renderer');
 
-const <%= className %> = require('../<%= className %>');
+const <%= className %> = require('.');
 
 describe('<%= className %>', () => {
   test('It renders', () => {
